src/app/requirements.txt
appdirs==1.4.4
bcrypt==3.2.0
black==20.8b1
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
ecdsa==0.14.1
fastapi==0.63.0
firebase-admin==4.5.2
flake8==3.8.4
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.27.1
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.36.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.36.1
h11==0.12.0
httplib2==0.19.0
idna==2.10
mccabe==0.6.1
msgpack==1.0.2
mypy-extensions==0.4.3
packaging==20.9
passlib==1.7.4
pathspec==0.8.1
proto-plus==1.14.2
protobuf==3.15.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.8
pyflakes==2.2.0
pyparsing==2.4.7
python-jose==3.2.0
python-multipart==0.0.5
pytz==2021.1
regex==2020.11.13
requests==2.25.1
rsa==4.7.2
six==1.15.0
starlette==0.13.6
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
uvicorn==0.13.4
